create: route for upload image to user and order

+++ src/usecases/images/delete/delete-images-usecase.ts
@@ -1,3 +1,7 @@
+import { IFileProvider } from "@/providers/StorageProvider/file-provider.interface"
+import { IStorageProvider } from "@/providers/StorageProvider/storage-provider.interface"
+import { IImageRepository } from "@/repositories/interfaces/interface-images-repository"
+import { AppError } from "@/usecases/errors/AppError"
 
 interface IRequestDeleteImages{
     id: string
@@ -5,7 +9,7 @@ interface IRequestDeleteImages{
 
 export class DeleteImageUseCase {
     constructor(
-        private imageRepository: IImagesRepository,
+        private imageRepository: IImageRepository,
         private storageProvider: IStorageProvider,
         private fileProvider: IFileProvider
         ) {}
@@ -21,13 +25,26 @@ export class DeleteImageUseCase {
             throw new AppError('Image not found', 404)
         }
 
-        // deletar imagens no firebase storage
-        await this.storageProvider.deleteFile(findImageExists.name, 'campings')
+        if(findImageExists.idUser){
+            // deletar imagens no firebase storage
+            await this.storageProvider.deleteFile(findImageExists.hashName as string, 'users')
 
-        // deletar imagen local no tmp
-        this.fileProvider.deleteFileTmp(findImageExists.name, 'campings')
+            // deletar imagen local no tmp
+            this.fileProvider.deleteFileTmp(findImageExists.hashName as string, 'users')
 
-        // deletar image pelo id
-        await this.imageRepository.deleteById(id)
+            // deletar image pelo id
+            await this.imageRepository.deleteById(id)
+        }
+
+        if(findImageExists.idOrder){
+            // deletar imagens no firebase storage
+            await this.storageProvider.deleteFile(findImageExists.hashName as string, 'orders')
+
+            // deletar imagen local no tmp
+            this.fileProvider.deleteFileTmp(findImageExists.hashName as string, 'orders')
+
+            // deletar image pelo id
+            await this.imageRepository.deleteById(id)
+        }
     }
 }
